+++ src/main/java/eu/gir/girsignals/signalbox/config/KSSignalConfig.java
@@ -41,15 +41,15 @@ public final class KSSignalConfig implements ISignalAutoconfig {
     @Override
     public void change(final int speed, final SignalTileEnity current, final SignalTileEnity next) {
         final HashMap<SEProperty, Object> values = new HashMap<>();
+        if (speed <= 16 && speed > 0) {
+            final ZS32 zs32 = ZS32.values()[ZS32.Z.ordinal() + speed];
+            values.put(SignalKS.ZS3, zs32);
+
+        }
 
         if (next != null) {
             current.getProperty(SignalKS.ZS3V).ifPresent(_u -> next.getProperty(SignalKS.ZS3)
                     .ifPresent(value -> current.setProperty(SignalKS.ZS3V, (ZS32) value)));
-            if (speed <= 16 && speed > 0) {
-                final ZS32 zs32 = ZS32.values()[ZS32.Z.ordinal() + speed];
-                current.getProperty(SignalKS.ZS3)
-                        .ifPresent(_u -> current.setProperty(SignalKS.ZS3, zs32));
-            }
             final boolean changes = next.getProperty(SignalKS.ZS3)
                     .filter(e -> ((ZS32) e).ordinal() > ZS32.Z.ordinal()
                             && (((ZS32) e).ordinal() - ZS32.Z.ordinal()) < speed)
@@ -58,7 +58,6 @@ public final class KSSignalConfig implements ISignalAutoconfig {
             final Optional opt = next.getProperty(SignalKS.STOPSIGNAL);
             final boolean stop = next.getProperty(SignalKS.MAINSIGNAL).filter(KSMain.HP0::equals)
                     .isPresent() || opt.filter(KS.HP0::equals).isPresent();
-
             final boolean hlstop = next.getProperty(SignalHL.STOPSIGNAL)
                     .filter(a -> STOP_HL.contains(a)).isPresent()
                     || next.getProperty(SignalHL.EXITSIGNAL).filter(d -> STOP_HL_EXIT.contains(d))
@@ -88,19 +87,32 @@ public final class KSSignalConfig implements ISignalAutoconfig {
                     || next.getProperty(SignalHV.HPHOME).filter(HPHome.HP0_ALTERNATE_RED::equals)
                             .isPresent();
             final Optional<ZS32> speedHV = (Optional<ZS32>) next.getProperty(SignalHV.ZS3);
+            final Optional<ZS32> speedHVplate = (Optional<ZS32>) next
+                    .getProperty(SignalHV.ZS3_PLATE);
             final boolean nexthl = next.getProperty(SignalHL.STOPSIGNAL).isPresent()
                     || next.getProperty(SignalHL.EXITSIGNAL).isPresent();
             final Optional<HLLightbar> nextlighbar = (Optional<HLLightbar>) next
                     .getProperty(SignalHL.LIGHTBAR);
+
             final Optional<ZS32> currentzs3v = (Optional<ZS32>) current.getProperty(SignalKS.ZS3V);
+            final Optional<ZS32> speedKSZS3plate = (Optional<ZS32>) next
+                    .getProperty(SignalKS.ZS3_PLATE);
+            final Optional<KS> currentks = (Optional<KS>) current.getProperty(SignalKS.STOPSIGNAL);
+            if (speedKSZS3plate.isPresent() && currentks.isPresent()
+                    && !currentks.filter(KS.HP0::equals).isPresent()) {
+                final ZS32 speednext = speedKSZS3plate.get();
+                values.put(SignalKS.ZS3V, speednext);
+                values.put(SignalKS.STOPSIGNAL, KS.KS1_BLINK);
+            }
 
             if (stop || hlstop) {
                 values.put(SignalKS.STOPSIGNAL, KS.KS2);
                 values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS2);
+                values.put(SignalKS.ZS3V, ZS32.OFF);
             } else if (changes) {
                 values.put(SignalKS.STOPSIGNAL, KS.KS1_BLINK);
                 values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1_BLINK);
-            } else {
+            } else if (!speedKSZS3plate.isPresent()) {
                 values.put(SignalKS.STOPSIGNAL, KS.KS1);
                 values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1);
             }
@@ -137,22 +149,26 @@ public final class KSSignalConfig implements ISignalAutoconfig {
                 values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1);
                 if (hv40 && currentzs3v.isPresent()) {
                     values.put(SignalKS.STOPSIGNAL, KS.KS1_BLINK);
+                    values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1_BLINK);
                     values.put(SignalKS.ZS3V, ZS32.Z4);
                 }
-                if (speedHV.isPresent() && currentzs3v.isPresent()) {
-                    final ZS32 speednext = speedHV.get();
+                if ((speedHV.isPresent() || speedHVplate.isPresent()) && currentzs3v.isPresent()) {
+                    final ZS32 speednext = speedHV.isPresent() ? speedHV.get() : speedHVplate.get();
                     final int zs32 = speednext.ordinal();
                     if (zs32 > 26 && zs32 <= 42 && zs32 != 30) {
                         values.put(SignalKS.STOPSIGNAL, KS.KS1_BLINK);
+                        values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1_BLINK);
                         values.put(SignalKS.ZS3V, speednext);
                     }
                     if (zs32 == 30) {
                         values.put(SignalKS.STOPSIGNAL, KS.KS1_BLINK);
+                        values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS1_BLINK);
                         values.put(SignalKS.ZS3V, ZS32.Z4);
                     }
                 }
             } else if (hvstop || hvstop2) {
                 values.put(SignalKS.STOPSIGNAL, KS.KS2);
+                values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS2);
                 values.put(SignalKS.DISTANTSIGNAL, KSDistant.KS2);
             }
         } else {
@@ -160,10 +176,10 @@ public final class KSSignalConfig implements ISignalAutoconfig {
             values.put(SignalKS.STOPSIGNAL, KS.KS2);
             values.put(SignalKS.MAINSIGNAL, KSMain.KS1);
             values.put(SignalKS.ZS2, ZS32.OFF);
-            values.put(SignalKS.ZS3, ZS32.OFF);
             values.put(SignalKS.ZS2V, ZS32.OFF);
             values.put(SignalKS.ZS3V, ZS32.OFF);
         }
+
         this.changeIfPresent(values, current);
     }